refactor: move tokens.ts to shared/ root
